+++ README.md
@@ -3,21 +3,21 @@ Weez Common
 
 * Install database of test
 ```
-Weez/ZendModelGenerator/demo/scripts/create_testzdmg_db.sql
+Weez/Generator/demo/scripts/create_testzdmg_db.sql
 
 ```
 
 * Setup config file (copy config.php.dist)
 
 ```
-Weez/ZendModelGenerator/config/config.php.dist
+Weez/Generator/config/config.php.dist
 
 ```
 
 * Generate Entity and EntityManager
 
 ```
-php weezcmd app:model-generator Weez/ZendModelGenerator/config/config.php testzdmg Weez\\ZendModelGenerator\\demo\\Core\\Model Weez/ZendModelGenerator/demo/Core/Model --tables-all --zfv 2
+php weezcmd app:model-generator Weez/Generator/config/config.php testzdmg Weez\\Generator\\demo\\Core\\Model Weez/Generator/demo/Core/Model --tables-all --zfv 2
 
 ```
 
@@ -31,4 +31,4 @@ php -S 0.0.0.0:8888
 
 * Lanch browser
 
-http://localhost:8888/Weez/ZendModelGenerator/demo
+http://localhost:8888/Weez/Generator/demo
