+++ Weez/Generator/Lib/ZendCode/EntityItem.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Weez\ZendModelGenerator\Lib\ZendCode;
+namespace Weez\Generator\Lib\ZendCode;
 
 use \Zend\Code\Generator\ClassGenerator;
 use Zend\Code\Generator\DocBlock\Tag\GenericTag;
