+++ Weez/Generator/demo/index.php
@@ -2,9 +2,9 @@
 
 use Pimple\Container;
 use Weez\Model\ModelFactory;
-use Weez\ZendModelGenerator\demo\Core\Model\Entity\Accounts;
-use Weez\ZendModelGenerator\demo\Core\Model\Entity\Bugs;
-use Weez\ZendModelGenerator\demo\Core\Model\Entity\User;
+use Weez\Generator\demo\Core\Model\Entity\Accounts;
+use Weez\Generator\demo\Core\Model\Entity\Bugs;
+use Weez\Generator\demo\Core\Model\Entity\User;
 use Zend\Db\Adapter\Adapter;
 
 require_once $_SERVER['DOCUMENT_ROOT'] . '/vendor/autoload.php';
@@ -27,7 +27,7 @@ $app['entity'] = function() use ($app) {
 };
 
 $timestart = microtime(true);
-$userManager = $app['entity']->get('\Weez\ZendModelGenerator\demo\Core\Model\Table\User');
+$userManager = $app['entity']->get('\Weez\Generator\demo\Core\Model\Table\User');
 $nbUser      = 10;
 echo "***********************************************" . "<br>";
 //Ajouter x users
